requirements.txt
alembic==1.10.4
amqp==5.1.1
apache-superset==2.1.0
apispec==5.2.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backoff==2.2.1
bcrypt==4.0.1
billiard==3.6.4.0
bleach==3.3.1
Brotli==1.0.9
cachelib==0.4.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
convertdate==2.4.0
cron-descriptor==1.2.35
croniter==1.3.14
cryptography==39.0.2
Deprecated==1.2.13
deprecation==2.1.0
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
Flask==2.1.3
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-Compress==1.13
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.0
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.0.0
Flask-WTF==1.0.1
func-timeout==4.3.5
geographiclib==2.0
geopy==2.3.0
graphlib-backport==1.0.3
gunicorn==20.1.0
h11==0.14.0
hashids==1.3.1
hijri-converter==2.3.1
holidays==0.17.2
humanize==4.6.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
korean-lunar-calendar==0.3.1
limits==3.4.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
mdurl==0.1.2
msgpack==1.0.5
mysqlclient==2.1.1
numpy==1.23.5
ordered-set==4.1.0
outcome==1.2.0
packaging==23.1
pandas==1.5.3
paramiko==3.1.0
parsedatetime==2.6
pgsanity==0.2.9
pkgutil_resolve_name==1.3.10
polyline==2.0.0
prison==0.2.1
prompt-toolkit==3.0.38
pyarrow==10.0.1
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
PyMeeus==0.5.12
PyMySQL==1.0.3
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-geohash==0.8.5
pytz==2023.3
PyYAML==6.0
redis==4.5.4
rich==13.3.5
selenium==4.9.0
simplejson==3.19.1
six==1.16.0
slack-sdk==3.21.3
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.48
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.3
sshtunnel==0.4.0
tabulate==0.8.10
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.2
wrapt==1.15.0
wsproto==1.2.0
WTForms==2.3.3
WTForms-JSON==0.3.5
XlsxWriter==3.0.9
zipp==3.15.0
